Rename modules and fix docs
