+++ src/dynamics/rigid_body/mass_properties/components/collider.rs
