+++ src/dynamics/rigid_body/mass_properties/components/computed.rs
